requirements.txt
langgraph
groq
streamlit
ipykernel
python-dotenv
langchain-groq
langgraph
langgraph-cli[inmem]